[WIP] memcpy.h -> fastcopy.c so as to be able to use it from other places
